+++ front-end/src/App.css
@@ -13,3 +13,7 @@ body {
   margin: 0;
   padding: 0;
 }
+
+#root {
+  min-height: 100vh;
+}
