+++ front-end/src/components/Login/About.module.css
@@ -1,0 +1,10 @@
+.about {
+  color: var(--text-color);
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  text-align: center;
+  gap: 15px;
+  max-width: 600px;
+  grid-area: about;
+}
